requirements.txt
pyautogui
PyYAML
label-studio
opencv-python
numpy
argparse
ultralytics
random
